finish all improvements

+++ backend/src/controllers/items/controller.ts
@@ -1,6 +1,9 @@
 import { NextFunction, Response, Request } from "express";
 import Item from "../../model/item";
 import Category from "../../model/category";
+import { Op } from "sequelize";
+import AppError from "../../errors/app-error";
+import { StatusCodes } from "http-status-codes";
 
 
 
@@ -18,6 +21,12 @@ export async function getAllItems(req: Request, res: Response, next: NextFunctio
 export async function getItemsPerCategory(req: Request<{ categoryId: string }>, res: Response, next: NextFunction) {
     try {
         const { categoryId } = req.params
+        
+        const category = await Category.findByPk(categoryId)
+        if (!category) {
+            throw new AppError(StatusCodes.NOT_FOUND, `Category with ID ${categoryId} not found`)
+        }
+
         const items = await Item.findAll({
             where: {categoryId},
             include: [Category]
@@ -32,7 +41,9 @@ export async function getItemsPerPrice(req: Request, res: Response, next: NextFu
     try {
         const price = Number(req.query.price)
         const items = await Item.findAll({
-            where: {price},
+            where: {price: {
+                [Op.lte]: price
+            }},
             include: [Category]
         })
         res.json(items)
@@ -43,9 +54,11 @@ export async function getItemsPerPrice(req: Request, res: Response, next: NextFu
 
 export async function getItemsPerIsRecycled(req: Request, res: Response, next: NextFunction) {
     try {
-        const isRecycled = req.query.isRecycled === 'true'
+        console.log(req.query.isRecycled)
+        const isRecycledParam = req.query.isRecycled ? 1 : 0
+        console.log(isRecycledParam)
         const items = await Item.findAll({
-            where: {isRecycled},
+            where: {isRecycled: isRecycledParam},
             include: [Category]
         })
         res.json(items)
@@ -63,9 +76,16 @@ export async function addItem(req: Request<{}, {}, {
     price: number,
     discount: number
 }>, res: Response, next: NextFunction) {
+
     try {
+        const categoryExists = await Category.findByPk(req.body.categoryId)
+        if (!categoryExists) {
+            throw new AppError(StatusCodes.BAD_REQUEST, `Cannot add item: Category with ID ${req.body.categoryId} not found`)
+        }
+        
         const newItem = await Item.create(req.body)
-        res.json(newItem)
+        res.status(StatusCodes.CREATED).json(newItem)
+
     } catch (error) {
         next(error)
     }
@@ -91,8 +111,14 @@ export async function editItem(req: Request<{ id: string}, {}, {
 export async function removeItem(req: Request<{ id: string }>, res: Response, next: NextFunction) {
     try {
         const {id} = req.params
+        const item = await Item.findByPk(id)
+
+        if (!item) {
+            throw new AppError(StatusCodes.NOT_FOUND, `Item with ID ${id} not found`)
+        }
+        
         await Item.destroy({where: {id}})
-        res.json({success: true})
+        res.status(StatusCodes.OK).json({success: true})
     } catch (error) {
         next(error)
     }
